requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
eth-account>=0.9.0
web3>=6.11.0
python-dotenv>=1.0.0
pydantic>=2.5.0
flask>=3.0.0
